Doc - Take whatever line endings and turn them into LF in generation (#5714)

+++ Documentation/Books/build.sh
@@ -404,11 +404,13 @@ function build-book()
     if ditaa --help > /dev/null; then
         echo "${STD_COLOR} - generating ditaa images${RESET}"
         find "${NAME}" -name "*.ditaa" | while IFS= read -r image; do
+            mkdir -p $(dirname "ppbooks/${image//ditaa/png}")
             ditaa "${image}" "ppbooks/${image//ditaa/png}"
         done
     else
         echo "${ERR_COLOR} - generating FAKE ditaa images - no ditaa installed${RESET}"
         find "${NAME}" -name "*.ditaa" | while IFS= read -r image; do
+            mkdir -p $(dirname "ppbooks/${image//ditaa/png}")
             cp "../../js/node/node_modules/mocha/images/error.png" \
                "ppbooks/${image//ditaa/png}"
         done
@@ -427,14 +429,10 @@ function build-book()
 
     (
         cd "ppbooks/${NAME}"
+        mkdir -p styles
         cp -a "../../${NAME}/styles/"* styles/
     )
     WD=$(pwd)
-    echo "${STD_COLOR} - copying images${RESET}"
-    find "${NAME}" -name "*.png" | while IFS= read -r pic; do
-        cd "${WD}/ppbooks"
-        cp "${WD}/${pic}" "${pic}"
-    done
 
     echo "${STD_COLOR} - generating MD-Files${RESET}"
     python ../Scripts/generateMdFiles.py \
@@ -495,8 +493,6 @@ function build-book()
     rm -f "./books/${NAME}/FOOTER.html"
     echo "${STD_COLOR} - deleting markdown files in output (gitbook 3.x bug)"
     find "./books/${NAME}/" -type f -name "*.md" -delete
-    echo "${STD_COLOR} - putting in deprecated items ${RESET}"
-    python ../Scripts/deprecated.py || exit 1
 
     book-check-markdown-leftovers "${NAME}"
 }
@@ -566,7 +562,7 @@ function check-docublocks()
         grep -v '.*#.*:.*' \
              >> /tmp/rawindoc.txt
 
-    sed  -e "s;.*ck ;;" -e "s;.*ne ;;" < /tmp/rawindoc.txt |sort -u > /tmp/indoc.txt
+    sed  -e "s;\r$;;" -e "s;.*ck ;;" -e "s;.*ne ;;" < /tmp/rawindoc.txt |sort -u > /tmp/indoc.txt
 
     set +e
     grep -R '^@startDocuBlock' ../DocuBlocks --include "*.md" |grep -v aardvark > /tmp/rawinprog.txt
@@ -587,7 +583,7 @@ function check-docublocks()
     set -e
     echo "Generated: startDocuBlockInline errorCodes">> /tmp/rawinprog.txt
 
-    sed -e "s;.*ck ;;" -e "s;.*ne ;;" < /tmp/rawinprog.txt  |sort > /tmp/inprog_raw.txt
+    sed -e "s;\r$;;" -e "s;.*ck ;;" -e "s;.*ne ;;" < /tmp/rawinprog.txt  |sort > /tmp/inprog_raw.txt
     sort -u < /tmp/inprog_raw.txt > /tmp/inprog.txt
 
     if test "$(wc -l < /tmp/inprog.txt)" -ne "$(wc -l < /tmp/inprog_raw.txt)"; then 
@@ -671,6 +667,10 @@ function build-dist-books()
 {
     set -x
     set -e
+    if test -z "${OUTPUT_DIR}"; then
+        echo "please specify --outputDir"
+        exit 1
+    fi
     rm -rf books ppbooks
     PIDFILE=/tmp/xvfb_20_0.pid
     if test "${isCygwin}" -eq 0 -a -z "${DISPLAY}"; then
